Make the test pass
I had to manually serialize the query set returned by the view and
manually parse the JSON response to compare against a manually
constructed dict. This seems less than ideal. I to find a way for
Django to automatically serialize QuerySets and Models and a way to
assert the returned result without so much manual parsing on my part.